chg: dev: "selectGroup" function moved
